renamed some files and added some css
