+++ api/database.json
@@ -11,18 +11,6 @@
       "phone_number": "000-111-9999",
       "address": "129 West 81st Street, Apartment 5A, New York, New York",
       "id": 2
-    },
-    {
-      "name": "austi",
-      "phone_number": "es3`",
-      "address": "asd",
-      "id": 3
-    },
-    {
-      "name": "Sadie",
-      "phone_number": "619",
-      "address": "75006  fake rd",
-      "id": 4
     }
   ]
 }
