+++ public/main.css
@@ -1,0 +1,15 @@
+.contact-card{
+  border: 1px solid dodgerblue;
+  width: 29%;
+  margin: 5px;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+#contact-list{
+  display: flex;
+  justify-content: center;
+  align-content: space-around;
+  flex-wrap: wrap;  
+}
